+++ tests/abi_test.go
@@ -15,8 +15,7 @@ import (
 	"github.com/yihuang/go-abi"
 )
 
-//go:generate go run ../cmd -var TestABI -output test.abi.go -prefix test -buildtag=!uint256
-//go:generate go run ../cmd -var TestABI -output test_uint256.abi.go -prefix test -buildtag=uint256 -uint256
+//go:generate go run ../cmd -var TestABI -output test.abi.go -prefix test
 
 // TestABI contains human-readable ABI definitions for testing
 var TestABI = []string{