+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Breaking Changes
+
+* (generator) [#25](https://github.com/yihuang/go-abi/pull/25) Removed `-uint256` flag and `generator.UseUint256(...)`, uint256 backing is now selected by `-tags uint256` at compile time, and flag from `//go:generate` directives is dropped.
+
 ### Bug Fixes
 
 ### Improvements